problems fixed
