+++ img_rec_module/img_20_3_meas.txt
@@ -1,43 +1,24 @@
 ksize:(5, 5), sigmax:0
-x: zero_crossing: -1, edge_converge: 228.11759868421052; 
-y: zero_crossing: 222.4569536423841, edge_converge: 224.60727272727274
-ksize:(5, 5), sigmax:1
-x: zero_crossing: -1, edge_converge: 228.12793522267205; 
-y: zero_crossing: 222.46504559270517, edge_converge: 224.5452380952381
+x: zero_crossing: 43.1578947368421, edge_converge: 27.907894736842106; 
+y: zero_crossing: 222.41739130434783, edge_converge: 224.8991379310345
 ksize:(7, 7), sigmax:0
-x: zero_crossing: -1, edge_converge: 222.77301587301588; 
-y: zero_crossing: 222.62559241706163, edge_converge: 224.3421052631579
-ksize:(7, 7), sigmax:1
-x: zero_crossing: -1, edge_converge: 228.11759868421052; 
-y: zero_crossing: 222.43730886850153, edge_converge: 224.5328282828283
+x: zero_crossing: -1, edge_converge: 222.0; 
+y: zero_crossing: 222.53191489361703, edge_converge: 228.6346153846154
 ksize:(9, 9), sigmax:0
-x: zero_crossing: -1, edge_converge: 228.31222222222223; 
-y: zero_crossing: 222.72103004291844, edge_converge: 208.65
-ksize:(9, 9), sigmax:1
-x: zero_crossing: -1, edge_converge: 228.11759868421052; 
-y: zero_crossing: 222.43730886850153, edge_converge: 224.5328282828283
-
-ksize:(5, 5), sigmax:0
-x: zero_crossing: 225.20987654320987, edge_converge: 219.04166666666669;
-y: zero_crossing: 222.6821917808219, edge_converge: 229.85833333333332
-ksize:(7, 7), sigmax:0
-x: zero_crossing: 224.8235294117647, edge_converge: 221.65238095238095;
-y: zero_crossing: 222.9, edge_converge: 224.62179487179486
-ksize:(9, 9), sigmax:0
-x: zero_crossing: 224.59375, edge_converge: 221.13636363636363;
-y: zero_crossing: 223.02487562189054, edge_converge: 225.125
+x: zero_crossing: -1, edge_converge: 225.0; 
+y: zero_crossing: 222.6338028169014, edge_converge: 213.0
 ksize:(11, 11), sigmax:0
-x: zero_crossing: 224.74444444444444, edge_converge: 221.13636363636363;
-y: zero_crossing: 223.28571428571428, edge_converge: 225.125
+x: zero_crossing: -1, edge_converge: -1; 
+y: zero_crossing: 222.75, edge_converge: 291.0
 ksize:(13, 13), sigmax:0
-x: zero_crossing: 224.7283950617284, edge_converge: 221.13636363636363;
-y: zero_crossing: 223.38888888888889, edge_converge: 225.125
+x: zero_crossing: -1, edge_converge: -1; 
+y: zero_crossing: 222.79166666666666, edge_converge: 200.0
 ksize:(15, 15), sigmax:0
-x: zero_crossing: 224.67441860465115, edge_converge: 221.13636363636363;
-y: zero_crossing: 223.3942857142857, edge_converge: 225.125
+x: zero_crossing: -1, edge_converge: -1; 
+y: zero_crossing: 222.7450980392157, edge_converge: 200.0
 ksize:(17, 17), sigmax:0
-x: zero_crossing: 224.67816091954023, edge_converge: 221.13636363636363;
-y: zero_crossing: 223.49101796407186, edge_converge: 225.125
+x: zero_crossing: -1, edge_converge: -1; 
+y: zero_crossing: 222.92105263157896, edge_converge: 200.0
 ksize:(19, 19), sigmax:0
-x: zero_crossing: 224.68493150684932, edge_converge: 221.13636363636363;
-y: zero_crossing: 223.58641975308643, edge_converge: 225.125
+x: zero_crossing: -1, edge_converge: -1; 
+y: zero_crossing: 222.9268292682927, edge_converge: 200.0
